tehtavat

+++ tehtavat/tehtava2.py
@@ -6,7 +6,7 @@ def nimi():
     print("Terve, " +nimi + "!")
 
 def ympyra():
-    sade = int(input("Mikä on ympyrän säde?"))
+    sade = float(input("Mikä on ympyrän säde?"))
     pintaala = 3.14 * (sade*sade)
     print("Ympyrän pinta-ala on ", pintaala)
